requirements.txt
Pillow
pytesseract
beautifulsoup4
requests
numpy
pandas
pytest
sphinx
opencv-python
onnxruntime
torchvision